Refactor the content widget
