+++ classes/admin/class-page.php
@@ -41,10 +41,8 @@ class Page {
 			\progress_planner()->get_widgets__suggested_tasks(),
 			\progress_planner()->get_widgets__todo(),
 			\progress_planner()->get_widgets__badge_streak(),
-			\progress_planner()->get_widgets__latest_badge(),
-			\progress_planner()->get_widgets__published_content_density(),
-			\progress_planner()->get_widgets__published_words(),
 			\progress_planner()->get_widgets__published_content(),
+			\progress_planner()->get_widgets__latest_badge(),
 			\progress_planner()->get_widgets__whats_new(),
 		];
