+++ classes/widgets/class-published-content.php
@@ -59,6 +59,34 @@ final class Published_Content extends \Progress_Planner\Widget {
 	 *
 	 * @return array The chart args.
 	 */
+	public function get_chart_args_content_density() {
+		return array_merge(
+			$this->get_chart_args(),
+			[
+				'count_callback' => [ $this, 'count_density' ],
+			]
+		);
+	}
+
+	/**
+	 * Get the chart args.
+	 *
+	 * @return array The chart args.
+	 */
+	public function get_chart_args_content_count() {
+		return array_merge(
+			$this->get_chart_args(),
+			[
+				// 'count_callback' => [ $this, 'count_words' ],
+			]
+		);
+	}
+
+	/**
+	 * Get the chart args.
+	 *
+	 * @return array The chart args.
+	 */
 	public function get_chart_args() {
 		return [
 			'query_params'   => [
@@ -95,5 +123,106 @@ final class Published_Content extends \Progress_Planner\Widget {
 					&& \in_array( $post->post_type, \progress_planner()->get_activities__content_helpers()->get_post_types_names(), true );
 			}
 		);
+	}
+
+	/**
+	 * Callback to count the words in the activities.
+	 *
+	 * @param \Progress_Planner\Activities\Content[] $activities The activities array.
+	 *
+	 * @return int
+	 */
+	public function count_words( $activities ) {
+		$words = 0;
+		foreach ( $activities as $activity ) {
+			if ( null === $activity->get_post() ) {
+				continue;
+			}
+			$words += \progress_planner()->get_activities__content_helpers()->get_word_count(
+				$activity->get_post()->post_content,
+				(int) $activity->data_id
+			);
+		}
+		return $words;
+	}
+
+	/**
+	 * Callback to count the density of the activities.
+	 *
+	 * Returns the average number of words per activity.
+	 *
+	 * @param \Progress_Planner\Activities\Content[] $activities The activities array.
+	 *
+	 * @return int
+	 */
+	public function count_density( $activities ) {
+		$words = $this->count_words( $activities );
+		$count = count( $activities );
+		return (int) round( $words / max( 1, $count ) );
+	}
+
+	/**
+	 * Get the density of all activities.
+	 *
+	 * @return int
+	 */
+	public function get_all_activities_density() {
+		// Get the all-time average.
+		static $density;
+		if ( null === $density ) {
+			$activities = $this->filter_activities(
+				\progress_planner()->get_query()->query_activities(
+					[
+						'category' => 'content',
+						'type'     => 'publish',
+					]
+				)
+			);
+			$density    = $this->count_density( $activities );
+		}
+		return $density;
+	}
+
+	/**
+	 * Get the weekly activities density.
+	 *
+	 * @return int
+	 */
+	public function get_weekly_activities_density() {
+		static $density;
+		if ( null === $density ) {
+			// Get the weekly average.
+			$density = $this->count_density(
+				\progress_planner()->get_query()->query_activities(
+					[
+						'category'   => 'content',
+						'type'       => 'publish',
+						'start_date' => new \DateTime( '-7 days' ),
+					]
+				)
+			);
+		}
+		return $density;
+	}
+
+	/**
+	 * Get the weekly words count.
+	 *
+	 * @return int The weekly words count.
+	 */
+	public function get_weekly_words() {
+		static $weekly_words;
+		if ( null === $weekly_words ) {
+			$weekly_words = $this->count_words(
+				\progress_planner()->get_query()->query_activities(
+					[
+						'category'   => 'content',
+						'type'       => 'publish',
+						'start_date' => new \DateTime( '-7 days' ),
+					]
+				)
+			);
+		}
+		return $weekly_words;
 	}
 }
