+++ views/page-widgets/published-content.php
@@ -47,7 +47,7 @@ $sum_weekly  = array_sum( $stats['weekly'] );
 	</p>
 </div>
 <div class="prpl-graph-wrapper">
-	<?php \progress_planner()->get_chart()->the_chart( $prpl_widget->get_chart_args() ); ?>
+	<?php \progress_planner()->get_chart()->the_chart( $prpl_widget->get_chart_args_content_count() ); ?>
 </div>
 <table>
 	<thead>
@@ -67,3 +67,66 @@ $sum_weekly  = array_sum( $stats['weekly'] );
 		<?php endforeach; ?>
 	</tbody>
 </table>
+
+<hr>
+
+<div class="prpl-counter-big-wrapper">
+	<span class="counter-big-number">
+		<?php echo \esc_html( \number_format_i18n( (int) $prpl_widget->get_weekly_activities_density() ) ); ?>
+	</span>
+	<span class="counter-big-text">
+		<?php echo \esc_html_e( 'content density', 'progress-planner' ); ?>
+	</span>
+</div>
+<p>
+	<?php
+	if ( 0 === $prpl_widget->get_weekly_activities_density() ) {
+		printf(
+			/* translators: %s: All-time average number. */
+			\esc_html__( 'Your overall content density average is %1$s.', 'progress-planner' ),
+			\esc_html( \number_format_i18n( $prpl_widget->get_all_activities_density() ) )
+		);
+	} else {
+		printf(
+			/* translators: %1$s: number of words/post published this week. %2$s: All-time average number. */
+			\esc_html__( 'Your content has an average density of %1$s words per post in the last 7 days. Your overall content density average is %2$s.', 'progress-planner' ),
+			\esc_html( \number_format_i18n( $prpl_widget->get_weekly_activities_density() ) ),
+			\esc_html( \number_format_i18n( $prpl_widget->get_all_activities_density() ) )
+		);
+	}
+	?>
+</p>
+<div class="prpl-graph-wrapper">
+	<?php \progress_planner()->get_chart()->the_chart( $prpl_widget->get_chart_args_content_density() ); ?>
+</div>
+
+<hr>
+
+<div class="prpl-counter-big-wrapper">
+	<span class="counter-big-number">
+		<?php echo \esc_html( \number_format_i18n( (int) $prpl_widget->get_weekly_words() ) ); ?>
+	</span>
+	<span class="counter-big-text">
+		<?php echo \esc_html_e( 'words', 'progress-planner' ); ?>
+	</span>
+</div>
+<p>
+	<?php if ( 0 === $prpl_widget->get_weekly_words() ) : ?>
+		<?php \esc_html_e( 'You didn\'t write last week. Let\'s get started!', 'progress-planner' ); ?>
+	<?php else : ?>
+		<?php
+		printf(
+			\esc_html(
+				/* translators: %1$s: number of posts published this week. %2$s: Total number of posts. */
+				\_n(
+					'Great job! You have written %1$s word in the past 7 days.',
+					'Great job! You have written %1$s words in the past 7 days.',
+					$prpl_widget->get_weekly_words(),
+					'progress-planner'
+				)
+			),
+			\esc_html( \number_format_i18n( $prpl_widget->get_weekly_words() ) ),
+		);
+		?>
+	<?php endif; ?>
+</p>
